Solved (Fix)

+++ src/main/java/web/controllers/UserController.java
@@ -1,6 +1,5 @@
 package web.controllers;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
@@ -16,7 +15,6 @@ public class UserController {
 
     private final UserService userService;
 
-    @Autowired
     public UserController(UserService userService) {
         this.userService = userService;
     }
@@ -51,7 +49,6 @@ public class UserController {
     public String find(@ModelAttribute("user") User user) {
         Long id = userService.findByEmail(user.getEmail());
         System.out.println("Find id: " + id);
-
         return id > 0L ? "redirect:/" + id : "/find";
     }
 
@@ -68,18 +65,20 @@ public class UserController {
     }
 
     @PatchMapping("/{id}")
-    public String update(@Valid @ModelAttribute("user") User user, BindingResult bindingResult
-            , @PathVariable("id") Long id) {
+    public String update(@Valid @ModelAttribute("user") User user, BindingResult bindingResult) {
+
         if (bindingResult.hasErrors()) {
             return "/edit";
         }
-        userService.update(id, user);
+        userService.update(user);
         return "redirect:/";
     }
 
     @DeleteMapping("/{id}/edit")
     public String delete(@PathVariable("id") Long id) {
-        userService.deleteById(id);
+        if (id != null) {
+            userService.deleteById(id);
+        }
         return "redirect:/";
     }
